requirements.txt
flask
werkzeug
flask-login
flask_sqlalchemy
pytest
pytest-cov
flask_wtf
faker

